refactor: localStorage 유틸 사용하여 getPosts 리팩토링
