+++ frontend/src/features/post/api/getPost.ts
@@ -1,0 +1,6 @@
+import { getPosts } from "./getPosts";
+
+export const getPost = (id: string) => {
+  const posts = getPosts();
+  return posts.find((p) => p.id === id);
+};
